requirements.txt
flask==2.0.2
exif
gunicorn
msgpack-python==1.0.0
pandas==1.1.5
yaml==0.2.5
tqdm==4.50
cudatoolkit==10.2
pytorch==1.6
torchvision==0.7
pytorch-lightning==1.0.1
shutil